feature: adding integration tests
