+++ zstandard_web/test/README.md
@@ -1,0 +1,5 @@
+## test
+
+This package uses integration tests for testing.
+
+See `example/README.md` for more info.
