+++ zstandard_web/test/tests_exist_elsewhere_test.dart
@@ -1,0 +1,12 @@
+// ignore_for_file: avoid_print
+
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('Tell the user where to find the real tests', () {
+    print('---');
+    print('This package uses integration_test for its tests.');
+    print('See `example/README.md` for more info.');
+    print('---');
+  });
+}
